pyproject.toml
[project]
name = "house-of-novels"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "httpx>=0.27.0",
    "websocket-client>=1.8.0",
    "moviepy>=2.0.0",
]
